// src/app/models/new_vendor.ts
export interface NewVendor {
    
    name: string;
    email: string;
    password1: string;
    password2: string;
    country: string
    address: string;
    phone: string;
    whatsapp_phone:string;
    longitude: number;
    latitude: number;
    source: string;
    device_id: string;
    gender: string;
    birthday: string;
    uuid: string;
  }
  