// src/app/models/vendeursrfu.ts
export interface Vendeursrfu {
    uuid: string;
    country: string;
    username: string;
    name: string;
    first_name: string;
    last_name: string;
    email: string;
    last_login: Date;
    address: string;
    member_since: number;
    id_pays: number;
    pays: object;
    user_photo: string;
    id_type: number;
    is_active: boolean;
    rating: number;
    telephone: number;
    phone: string;
    whatsapp_phone: string;
    ads_count: number;
    acquisition_channel;
  }
  